SYSLINK: procmgr - adjust ducati memory map for 512MB
Ducati MMU programming is adjusted to use the appropriate physical
address from the now available 512MB of RAM. Ducati uses a total of
49MB set aside from boot args. The new PA starts from 0x9CF00000
(512MB-49MB + 0x80000000). The Ducati IPC region is programmed
first so that the PA aligns with the VA for the rest of the regions
for programming the 16MB TLB entries.

Signed-off-by: Suman Anna <s-anna@ti.com>