OMAP: DSS2: adjust YUV overlay width to be even
An overlay in YUV mode has to have an even input width, because data for
each pixel is divided between two adjacent pixels.

The algorithm handling manual update overlay adjusting may adjust the
overlay width to be odd.

This patch adds a check for that situation, and makes the width even.
The width is increased by one if it is possible (the unadjusted input
width is larger than the width), and decreased by one if increasing is
not possible.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@nokia.com>