Fix incorrect rotation for 90 and 270 degrees
The exisiting code rotated an image by 270 when asked for 90 degree
and vice vera, this is because tiler rotation rotates the container
while v4l2 rotates the resultant image from the container.The fix is
to pass the opposite rotation to tiler api.

Some unnecessary tiler api calls have been removed too.

Signed-off-by: Archit Taneja <archit@ti.com>