OMAP: Introduce device specific set rate and get rate in device opp structures.
This patch extends the device opp structure to contain
pointers to scale the operating rate of the
device and to retrieve the operating rate of the device.
This patch also adds the three new APIs in the opp layer
namely opp_set_rate that can be called to set a new operating
rate for a device, opp_get_rate that can be called to retrieve
the operating frequency for a device and opp_populate_rate_fns
to populte the device specific set_rate and get_rate API's.
The opp_set_rate and opp_get_rate does some routine error
checks and finally calls into the device specific set_rate
and get_rate APIs populated through opp_populate_rate_fns.

Signed-off-by: Thara Gopinath <thara@ti.com>