SYSLINK:IPU-PM: workqueue implementation for ipu_pm
Introducing the use of a workqueue to manage the request
and release of the resources using the ipu_pm module.
The implementation is using a FIFO to store the messages
(proc_id and payload) in the CB and enqueueing the work,
currently only one, then the work function is calling the
particular API to req/rel a resource.

Due to this implementation also some global variables were
changed to local.

FIFO depth is 16 but configurable.
Request and release API are now separated.

This implementation is based in the patch provided by
Fernando Guzman Lugo.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>
Signed-off-by: Fernando Guzman Lugo <x0095840@ti.com>