usb: ehci: omap: Add OMAP4 support
Update the ehci-omap driver to add support for OMAP4.

The OMAP4 functional clocks are enabled.The OMAP4 has two ports.
The funtional clock for each port is driven by external 60 mhz clocks.
Depending on either PHY mode or TLL mode of controller, the TLL clocks
are enabled and configure the two TLL channels.
The controller is set in No idle and stand by mode.

Signed-off-by: Keshava Munegowda <keshava_mgowda@ti.com>
Signed-off-by: Anand Gadiyar <gadiyar@ti.com>