omap4: cpufreq: Coordinate CPU0 and CPU1 since they share voltage and clock
In OMAP4, both processors share the same voltage and the same clock,
but have dedicated power domains. So both cores needs to be scaled
together and hence needs software co-ordination.

This patch updtaes the system "loops_per_jiffy" for the clock speed
change. Note that loops_per_jiffy is not updated on SMP systems in
cpufreq driver. So, update the per-CPU loops_per_jiffy value
on frequency transition.

Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Signed-off-by: Thara Gopinath <thara@ti.com>