cpufreq: helpers for higher & lower frequencies
Two new functions for getting the next highest and next lowest
frequencies in the cpufreq table, based upon a supplied frequency in
kHz.

This is useful for cpufreq governors that do not target new frequencies
based on a percentage or a pre-determined value, but instead access the
cpufreq table directly to determine the next target frequency.

Signed-off-by: Mike Turquette <mturquette@ti.com>

whitespace