OMAP4: Clock: add MN Bypass data to DPLL structs
OMAP4 DPLLs support a new type of bypass called "MN Bypass", which allows MN
dividers to be used on a bypassed DPLL.  Add this flag to supported DPLLs and
include the mask to check on MN Bypass status.

Also create a helper function to put DPLLs into MN Bypass.  This will be
useful in the future once existing set_rate functions are intelligent enough
to detect MN Bypass mode and change the MN dividers without trying to relock
the DPLL.

Change-Id: I55c9e3f41059c5bc33e7a751da75638506e63b0b
Signed-off-by: Mike Turquette <mturquette@ti.com>