OMAP4: OPP: add OPP_LP data for DPLL cascading
DPLL cascading is a unique low power mode in which DPLL_ABE is driven by
SYS_32K_CK, DPLL_CORE is driven by DPLL_ABE and all other DPLLs are driven
by DPLL_CORE.  This allows SYS_CK to be gated and the entire system runs at
lower clock frequencies while consuming less power.

In this configuration not all clock frequencies for each device are
supported.  Some devices simply should not run in this mode, others at
reduced rate and a special few are unaffected.  Consult the PRCM functional
spec for more detailed info.

The OPP data now has an OPP_LP entry for each device in the table.  These
OPPs are disabled by default and are mutually exclusive with the other OPPs:
only during DPLL cascading mode will they be enabled while all of the
"normal operation" OPPs are disabled.  When leaving DPLL cascading mode the
OPP_LPs are again disabled and the "normal operation" are enabled.  This
prevents the DVFS framework from exploding and respects the current
limitations of the system.  Specifically, DVFS transitions will not occur
and clock frequencies will be reported properly in the clock framework.

Change-Id: Icac891c166dc30e44750c029773fda656004bb92
Signed-off-by: Mike Turquette <mturquette@ti.com>