OMAP: Voltage: add ABB structures and data
OMAP3630 and OMAP4 support an Adaptive Body-Bias LDO on some voltage domains
which effects voltage supplied to PMOS back-gates.  These LDOs have
different operating modes which correspond to the voltage of each VDD.

This patch introduces the data structures needed to represent the ABB LDOs
in the voltage layer, and populates the appropriate data for 3630 and OMAP4.
Note that OMAP34XX does not have this LDO, so any data for that chip has
been initialized as zero.

Signed-off-by: Mike Turquette <mturquette@ti.com>