OMAP: Voltage: Adaptive Body-Bias handlers
Introduce voltage transition notification handlers for Adaptive Body-Bias
LDOs.  There is an ABB LDO for VDD_MPU on OMAP3630 and an ABB_LDO on VDD_MPU
and VDD_IVA on OMAP4430.

All of these LDOs are handled similary.  Initial configuration is to enable
the possibility of going into Forward Body-Bias (which boosts voltage at
high OPPs).  This feature was designed for weak silicon, and eFuse values
exist to control whether or not this feature should be turned on.  However
recommendations from hardware folks always say to leave it on regardless of
eFuse values, so we don't bother checking them.  For all other OPPs the LDO
is in bypass and will follow the voltage of it's corresponding VDD_xxx.
Reverse Body-Bias exists but we never use this in practice (for saving power
on strongly characterised silicon).

Upon a DVFS transition the notifiers handle the sequencing of voltage
scaling and ABB LDO transitions.  When moving to a higher OPP that needs
FBB, raise voltage first and then enable FBB.  When moving down to an OPP
that bypasses ABB, first bypass the LDO then lower voltage.

Signed-off-by: Mike Turquette <mturquette@ti.com>