OMAP4: OPP: struct data for DPLL cascading
Introduce OPPs for DPLL cascading.  These OPPs are disabled by default, but
will be enabled by when entering DPLL cascading, and disabled when leaving
that mode.

This makes use of the "voltage" trick, where the voltage can be listed as
"928000" in the OPP table to differentiate it from the OPP50 voltages, but
it will really be 930000 on the device.

Change-Id: I5c83fcacebba45badd647ab0784f31a9bd6ed4db
Signed-off-by: Mike Turquette <mturquette@ti.com>