OMAP: voltage: search for disabled OPPs by voltage
opp_find_freq_exact allows searching for OPPs that match the supplied
frequency, whether they are enabled or not.  This feature is useful when
searching by for OPPs by voltage also.

In this patch opp_find_voltage is modified to make use of this, and the one
use of the API in the DVFS code is updated to only search for enabled OPPs.

Change-Id: Iba9035c2c992439bc9d5ea588b987559aef82917
Signed-off-by: Mike Turquette <mturquette@ti.com>