OMAP3+: ABB: Adaptive Body-Bias structures & data
Due to voltage domain trimming and silicon characterstics some silicon
may experience instability when operating at a very high voltage.  To
compensate for this an Adaptive Body-Bias LDO exists, starting with
OMAP3630.  The purpose of this LDO is to provide a performance boost
when running at high OPPs.  The ABB LDO is said to be operating in
"Forward Body-Bias" in this mode.  At OPPs operating within a nominal
voltage range the ABB LDO is bypassed.

This patch introduces the data structures needed to represent the ABB
LDOs in the voltage layer, and populates the appropriate data for 3630
and OMAP4.  Note that OMAP34XX does not have this LDO, so any data for
that chip has been initialized as zero.

Signed-off-by: Mike Turquette <mturquette@ti.com>