OMAP3+: OPP: add ABB data to OPP tables
The operating mode of the Adaptive Body-Bias LDO maps directly to the
voltage of its voltage domain.  The two modes supported are bypass and
Forward Body-Bias (FBB).

This patch models this relationship by adding an abb_type paramter to
struct omap_volt_data and populates this type in the 3630, 4430 and 4460
OPP tables.

NOMINAL_OPP causes the ABB LDO to be in bypass at that specific voltage.
FAST_OPP unsurprisingly causes the LDO to operate in Forward Body-Bias
mode.

Two details to note:
1) 3430 does not have an ABB LDO at all, so it's abb_type values are
initialised to zero.

2) 44xx silicon does have an ABB LDO on VDD_IVA, but it is recommended
to leave the LDO in bypass at all OPPs.  As such all abb_type entries
are marked as NOMINAL_OPP.

Signed-off-by: Mike Turquette <mturquette@ti.com>